+++ wire-ios-sync-engine/Tests/Source/SessionManager/SessionManagerTests.swift
@@ -457,114 +457,6 @@ extension IntegrationTest {
 
 // MARK: - Account Deletion
 
-class SessionManagerTests_EncryptionAtRestMigration: IntegrationTest {
-
-    override var useInMemoryStore: Bool {
-        return false
-    }
-
-    override func setUp() {
-        DeveloperFlag.storage = UserDefaults(suiteName: UUID().uuidString)!
-        var flag = DeveloperFlag.proteusViaCoreCrypto
-        flag.isOn = false
-
-        super.setUp()
-
-        createSelfUserAndConversation()
-        createExtraUsersAndConversations()
-    }
-
-    override func tearDown() {
-        DeveloperFlag.storage = .standard
-        super.tearDown()
-    }
-
-    // @SF.Storage @TSFI.UserInterface @S0.1 @S0.2
-    func testThatDatabaseIsMigrated_WhenEncryptionAtRestIsEnabled() throws {
-        // given
-        XCTAssertTrue(login())
-        var session = try XCTUnwrap(userSession)
-        XCTAssertFalse(session.encryptMessagesAtRest)
-
-        let expectedText = "Hello World"
-        session.perform({
-            let groupConversation = self.conversation(for: self.groupConversation)
-            try! groupConversation?.appendText(content: expectedText)
-        })
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // when
-        try session.setEncryptionAtRest(enabled: true)
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // then
-        session = try XCTUnwrap(userSession)
-        XCTAssertTrue(session.encryptMessagesAtRest)
-
-        try session.unlockDatabase(with: LAContext())
-        let groupConversation = self.conversation(for: self.groupConversation)
-        let clientMessage = groupConversation?.lastMessage as? ZMClientMessage
-        XCTAssertEqual(clientMessage?.messageText, expectedText)
-    }
-
-    // @SF.Storage @TSFI.UserInterface @S0.1 @S0.2 
-    func testThatDatabaseIsMigrated_WhenEncryptionAtRestIsDisabled() throws {
-        // given
-        XCTAssertTrue(login())
-        var session = try XCTUnwrap(userSession)
-
-        let expectedText = "Hello World"
-
-        try session.setEncryptionAtRest(enabled: true, skipMigration: true)
-        XCTAssertTrue(session.encryptMessagesAtRest)
-
-        session.perform({
-            let groupConversation = self.conversation(for: self.groupConversation)
-            try! groupConversation?.appendText(content: expectedText)
-        })
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // when
-        try session.setEncryptionAtRest(enabled: false)
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // then
-        session = try XCTUnwrap(userSession)
-        XCTAssertFalse(session.encryptMessagesAtRest)
-
-        let groupConversation = self.conversation(for: self.groupConversation)
-        let clientMessage = groupConversation?.lastMessage as? ZMClientMessage
-        XCTAssertEqual(clientMessage?.messageText, expectedText)
-    }
-
-}
-
-class SessionManagerTests_EncryptionAtRestIsEnabledByDefault_Option: IntegrationTest {
-
-    override func setUp() {
-        super.setUp()
-        createSelfUserAndConversation()
-    }
-
-    override var useInMemoryStore: Bool {
-        return false
-    }
-
-    override var sessionManagerConfiguration: SessionManagerConfiguration {
-        return SessionManagerConfiguration(encryptionAtRestIsEnabledByDefault: true)
-    }
-
-    // @SF.Storage @TSFI.UserInterface @S0.1 @S0.2
-    func testThatEncryptionAtRestIsEnabled_OnActiveUserSession() {
-        // given
-        XCTAssertTrue(login())
-
-        // then
-        XCTAssertTrue(sessionManager?.activeUserSession?.encryptMessagesAtRest == true)
-    }
-
-}
-
 class SessionManagerTests_PasswordVerificationFailure_With_DeleteAccountAfterThreshold: IntegrationTest {
     private var threshold: Int? = 2
     override func setUp() {
